requirements.txt
boto3==1.9.134
botocore==1.12.134
Click==7.0
dnspython==1.16.0
docutils==0.14
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
pymongo==3.7.2
python-dateutil==2.8.0
python-dotenv==0.10.1
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.2
